Refactored code to separate files and applied enumerate()
